Define QStringLiteral when undefined (e.g. Qt 4)
This will allow us Qt 5 style coding. QLatin1String replaced everywhere where it
works for Qt 4 and 5, namely: the direct literal passing.

Note that the "standard" typo has also been fixed in this commit as the line had
to change anyway, so it is not much of an additional noise.

Change-Id: I8b7e4fe5f337441000bd3d8f58db528fdd0e175b
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>