Eliminate the "canAppendToList" variable usage
There are problems with this variable:

1) It should have been called "isSerialPortValid" or something like that to be
more precise abouts its functionality.

2) It should have had false value by default since we do not assume everything
is correct by default, and exclude certain ones, but the other way around,
logically.

3) Most importantly: It makes the code unnecessarily longer since the loop can
be continued with the next iteration without the need to insert an invalid
serial port and an interim variable.

Tested the cenumerator example on Linux with udev using Qt 4 and then 5. There
are no regressions found.

Change-Id: I4f6bdce402f03eac696c0d2ce9448e59baedf56b
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>