Simplify the standard baudrate getter by dropping the foreach
1f1ca7490fa6e2231f396e564dacbbd8d0e10b29 rewrote this part of the code, but the
needless logic has not be dropped respectively. It was necessary before the
rewrite since the raw qint32* type had to be converted to QList<qint32>, but
since we have a more Qt'ish data storage up front, we do not need to iterate
through the list since it is ready-made for us.

Change-Id: Iab5d4b0d48de3fc27fb41a4694f044bd58e760c4
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>