Remove unused variables
* actualWriteBufferSize variable - it is a wrong old heritage related with
the implementation of a code with the analogy from the the QWindowsPipeReader
(where actualReadBufferSize is present).

* acyncWritePosition variable - it is a old heritage when the write notifier
could operate with a multiple writing operations at the same time. For this
purpose it is necessary to know the following write position in the ring buffer.
But currently all write operations happen sequentially, i.e. forbidden to carry
out startAsyncWrite() several times without completeAsyncWrite() completion.

Change-Id: I25b665b638649e6002fb194babfaa73f5ec8a6fa
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>