Use QString for device properties
The device registry property of the SPDRP_HARDWAREID is not
used any more. This property was used earlier for parsing of
the VID/PID and had the type of the REG_MULTI_SZ which was
interpreted as QStringList, see commit:

d8dc10efb1714dcfafa6a08e107fd31fc1e3ce0e

The current code is using only string based concept, hence
changing the interface to QString.

Tested on Windows 8 with on-board and virtual com0com serial
ports using the cenumerator example.

Change-Id: I31e4f85f1b145021c1b34b81ba890553604a0531
Reviewed-by: Laszlo Papp <lpapp@kde.org>
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>