Remove the noisy comments from the terminal example
If anything, this should be made clear in a different way, like proper variable
names, inline convenience functions for naming, example documentation extension,
and so on.

This is also consistent with the other examples not using "water can" comments.

However, these comments mostly add noise to the code without much benefit as far
as I can tell.

Change-Id: I9c4121d91ca2671d44c08cf986232422f136a785
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>