Move the WinCE implementation into separate module
WinCE implementation was mixed a common code with the Win32
implementation that caused some problems with maintenance.

More correct decision is moving a common code which was used
in *win.cpp into *wince.cpp module without modifications.

Now the Win32 and the WinCE implementation are independent
from each other.

Tested build with Qt5 on "win32-msvc2012" and then with Qt4
on "wincewm50pocket-msvc2008" configurations.

Change-Id: I63f687468beffa9a75b534a4fbe536b854b12210
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>