Move out processing of ResoureError from processIoErrors()
Handling of the ResourceError is moved directly to
_q_completeAsyncCommunication() method.

Reasons:

* The processIoErrors() should handle only the
EV_ERR event.

* The handling of ResourceError in the processIoErrors()
it is the old heritage when the event processing logic
was into the notifiers.

Also the method processIoErrors() is renamed in favor to
handleLineStatusErrors() for readability, according to
documentation on EV_ERR mask:
http://msdn.microsoft.com/en-us/library/windows/desktop/aa363479%28v=vs.85%29.aspx

Tested build on Windows 7/8 using Qt4 and then Qt5.

Change-Id: I6dc331933155f1e182b21f40885e47d574c4ed9f
Reviewed-by: Dyami Caliri <dyami@dragonframe.com>
Reviewed-by: Rafael Roquetto <rafael.roquetto@kdab.com>
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>