Fix the error message for TimeoutError on Windows
At timeouts from the waitForX() methods are set a wrong
error string. Thing in that the WAIT_TIMEOUT constant is
not a system error, therefore call of GetLastError()
returns a code of a system error which was set by someone
earlier.

Thus, it is necessary to extract text representation of
an error directly from the WAIT_TIMEOUT constant without
relying on GetLastError().

Tested on Windows 7/8 using Qt4 and then Qt5.

Change-Id: Id7fdc71c1b64c39eed658148bb1b66ac8806f119
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Dyami Caliri <dyami@dragonframe.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>