Simplify of the availablePortsByUdev()
There is no sense to use udev_device_get_subsystem() for comparing with
the predetermined names of subsystems (like "pnp", "usb" and so on).
This idea was preliminary and required further revising after
accumulation of some statistic.

The main ideas in favor of refactoring are given below:

* It seems that reasonable to use the "ID_MODEL", "ID_VENDOR" and others
identifiers to query of properties for all types of serial ports. We
lose nothing, because in the worst case the getUdevPropertyValue()
function will simply return an empty string, as before.

* Earlier we excluded from the list all ports which have a subsystem
with the "platform" name. Practice showed that for these devices is used
the serial8250 driver. Therefore it is reasonable to exclude this type
of devices by the name of drivers instead of by the name of subsystems.

Also are made related changes:

* Is added the resolving of chars for the new udev_device_get_driver()
function

* The functions of getting properties are wrapped for reduction of
length up to 100 characters.

Tested on ArchLinux 64-bit with the on-board, PL2303 serial devices
using Qt4 and then Qt5.

Tested build on Android x86 using Qt5.

Change-Id: I1a66164ca1893180e1ed97524cff98b4a933a63b
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>